Added the ability to pin discussions
I migrated a lot of the pinned functionality from the the forum plugin.
A lot of changes that were made were moved over and adjusted to be compliant with the hsuforum.
Changes that i did myslef were updated to the sort class by adding a default value to the return, i also added a new image for the pin.
There was a update to the renderer to display the pin.